chore: Output colors conditionally
Do not output colors if the output is not tty (broke with git
wrappers that do not support colors).

Do not output colors if NO_COLOR is set.

More elegant solution would be to create some format method that
would apply the colors appropriately (think '[red]text[/]'), but
this should be good enough for now.