feat(sllm): Add --start and --stop flags
When 'ramalama serve' is running in background, it provides not just an API,
but also surprisingly good llama.cpp Web UI, which can be used to discuss
issues in plain text. The '--start' flag makes it a feature, and lets the
user start the server without having to run some of the specialized tools.

The '--stop' flag is a way to shut down ramalama without having to wait for
the timeout.