feat(sllm-code): Coding assistant
A new sllm family tool adds a prompt to implement methods based on
the method signature. The prompt is not perfect yet, but it is a
start; LLMs are really good at figuring out which language you are
using, so pasting in the method signature and a documentation works.

Later, this could also be expanded to '--explain' or others. By
default, we always open $EDITOR to allow you to write or format
the input properly.