fix(sllm): Update default image name
Ramalama (of version 0.9, at least), names the models differently than it
used to do. The default model is now called 'ollama://llama3.2/llama3.2:3b'.

This patch ensures we can properly display the size of the used model in
'--status'.

It is not a perfect solution, as user-supplied input may break this, but I don't
see any other simple fix for it now.