chore: Switch to Ollama
Ramalama has been suffering from various regressions since sllm
was written. Sometimes it doesn't work at all, other times it
doesn't return the output properly (e.g. it doesn't sanitize the
end tokens out of the response).

Since Ollama publishes an official container, we can have +- the
same level of security and features. And the fact that we need
one less tool installed on the system is a plus.